fix: 응답객체 DTO 수정

+++ src/main/java/com/yunbaek/blogsearchapplication/client/dto/kakao/Meta.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class Meta {
 
-	private Integer totalCount;
+	private Long totalCount;
 
 	private Integer pageableCount;
